Abstracted out some functions from the main app file
